LayoutTests/fast/forms/input-valueasdate-date-expected.txt

Tests for .valueAsDate with <input type=date>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS valueAsDateFor("") is null
PASS valueAsDateFor("1969-12-31").getTime() is Date.UTC(1969, 11, 31)
PASS valueAsDateFor("1970-01-01").getTime() is Date.UTC(1970, 0, 1)
PASS valueAsDateFor("2009-12-22").getTime() is Date.UTC(2009, 11, 22)
PASS successfullyParsed is true

TEST COMPLETE
